+++ views/partials/head.ejs
@@ -12,6 +12,8 @@
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
 <script type="text/javascript" src="/assets/js/post.js"></script>
 <script type="text/javascript" src="/assets/js/results.js"></script>
+<script type="text/javascript" src="/assets/js/restModal.js"></script>
+<script type="text/javascript" src="/assets/js/eventModal.js"></script>
 
 <style>
     body    { padding-top:50px; }
