+++ views/partials/header.ejs
@@ -9,19 +9,26 @@
 
     <div class="container-fluid">
     
-        <div class="navbar-header">
-            <a class="navbar-brand" href="/">
-                <span class="glyphicon glyphicon glyphicon-glass"></span>
-            </a>           
-
-            <ul class="nav navbar-nav">
-                <li><a href="/results">See Listings</a></li>
-                <li><a href="/post">Post Listings</a></li>
-                <li><a href="#">Login</a></li>                
-            </ul>
+        <!--<div class="navbar-header">-->
+            <div class="container" style="margin-left: 0px; margin-right: 0px; width: 100%;">
+                <div class="row">
+                    <div class="col-md-4" id="nav-paths">
+                        <a class="navbar-brand" href="/">
+                            <span class="glyphicon glyphicon glyphicon-glass"></span>
+                        </a>
+                        <a href="/post">Post Listings</a>        
+                    </div>
+                    <div class="col-md-4" id="nav-logo">
+                        <h1 style="margin-top: 10px;">Date Night</h1>
+                    </div>
+                    <div class="col-md-4" id="nav-login" style="text-align: right;">
+                        <a href="#" style="text-align: right; padding-left: 90%">Login</a>
+                    </div>
+                </div>
+            </div>
 
             
-        </div>
+        <!--</div>-->
     </div>
 </nav>
