+++ views/partials/postform.ejs
@@ -92,8 +92,8 @@
     <div class="form-group">
       <label class="col-md-4 control-label" for="textarea">Description</label>
       <div class="col-md-4">                     
-        <textarea class="form-control" id="textarea" name="textarea" style="height: 100px; width: 400px;"></textarea>
-        <input type="submit" value="Submit" id="submit">
+        <textarea class="form-control" id="eventDescription" name="textarea" style="height: 100px; width: 400px;"></textarea>
+        <input type="submit" value="Submit" id="submit" class="eventSubmit">
       </div>
     </div>
